Add the burstRead() method.
This can be a much faster method to read a length of data,
especially if you've configured the device FIFO and want to read it
all. You could do it by reading the FIFO_DATA_OUT_L and
FIFO_DATA_OUT_H registers manually, but this introduces a lot of
overhead - the burstRead() method is 2-3 times faster.